requirements.txt
fastapi
uvicorn
pillow
torch
opencv-python
